Structure rebuild, Added file manager, file classes responsible for storing data extracted frrom file and providing it to code widget

+++ core/HELPER/CONSTANTS.h
@@ -12,17 +12,17 @@
 
 namespace Constants{
     // --Window & layout
-    inline constexpr size_t WINDOW_WIDTH = 1260;
-    inline constexpr size_t WINDOW_HEIGHT = 720;
-    inline constexpr int FPS = 90;
-    inline constexpr int DELAY = 1000/FPS;
+    inline constexpr uint32_t WINDOW_WIDTH = 1260;
+    inline constexpr uint32_t WINDOW_HEIGHT = 720;
+    inline constexpr uint32_t FPS = 90;
+    inline constexpr uint32_t DELAY = 1000/FPS;
 
     // --Numerator plugin
-    inline constexpr size_t NUMERATION_WIDTH = 45;
-    inline constexpr size_t NUMERATION_X_OFFSET = 125;
-    inline constexpr size_t CODE_LINES_X_OFFSET = NUMERATION_X_OFFSET + NUMERATION_WIDTH;
-    inline constexpr size_t CODE_LINES_Y_OFFSET = 50;
-    inline constexpr size_t CODE_VIEWPORT_HEIGHT = WINDOW_HEIGHT - CODE_LINES_Y_OFFSET;
+    inline constexpr uint32_t NUMERATION_WIDTH = 45;
+    inline constexpr uint32_t NUMERATION_X_OFFSET = 125;
+    inline constexpr uint32_t CODE_LINES_X_OFFSET = NUMERATION_X_OFFSET + NUMERATION_WIDTH;
+    inline constexpr uint32_t CODE_LINES_Y_OFFSET = 50;
+    inline constexpr uint32_t CODE_VIEWPORT_HEIGHT = WINDOW_HEIGHT - CODE_LINES_Y_OFFSET;
 
 
     // --- Brushes ---
